Actualizar

+++ simwebpluscar/trunk/backend/models/aaee/declaracion/DeclaracionBaseForm.php
@@ -234,7 +234,7 @@
     			 'operator' => '>=', 'on' => 'definitiva',
     			 'message' => Yii::t('backend', '{attribute} must be no less that ' . self::attributeLabels()['monto_minimo'])],
     			[['monto_new', 'monto_v', 'monto_minimo'],
-    			  'double', 'message' => Yii::t('backend', '{attribute} must be decimal.')],
+    			  'float', 'message' => Yii::t('backend', '{attribute} must be decimal.')],
     			['monto_new', 'compare', 'compareValue' => 0, 'operator' => '>='],
 
 	        ];
